Moved TILE to its own class file.
This resolves the awkward import situation.
Also added a not clearly explaining the Tile/TILE situation.
